Removed some warnings

+++ ParticleToy/Particle.cpp
@@ -19,9 +19,9 @@ void Particle::draw() {
     const double h = 0.03;
     glColor3f(1.f, 1.f, 1.f);
     glBegin(GL_QUADS);
-    glVertex2f(m_Position[0] - h / 2.0, m_Position[1] - h / 2.0);
-    glVertex2f(m_Position[0] + h / 2.0, m_Position[1] - h / 2.0);
-    glVertex2f(m_Position[0] + h / 2.0, m_Position[1] + h / 2.0);
-    glVertex2f(m_Position[0] - h / 2.0, m_Position[1] + h / 2.0);
+    glVertex2f((GLfloat) (m_Position[0] - h / 2.0), (GLfloat) (m_Position[1] - h / 2.0));
+    glVertex2f((GLfloat) (m_Position[0] + h / 2.0), (GLfloat) (m_Position[1] - h / 2.0));
+    glVertex2f((GLfloat) (m_Position[0] + h / 2.0), (GLfloat) (m_Position[1] + h / 2.0));
+    glVertex2f((GLfloat) (m_Position[0] - h / 2.0), (GLfloat) (m_Position[1] + h / 2.0));
     glEnd();
 }
